third_party/blink/web_tests/http/tests/navigation/parsed-iframe-dynamic-form-back-entry-user-gesture.html

<html>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.setPopupBlockingEnabled(false);
    testRunner.dumpBackForwardList();
    testRunner.dumpAsText();
}

function runTest()
{
    if (!window.eventSender)
        return;
    var i = document.getElementById("submitform");
    eventSender.mouseMoveTo(i.offsetLeft + 2, i.offsetTop + 2);
    eventSender.mouseDown();
    eventSender.mouseUp();
}

function submit()
{
    window.open("resources/form-submit-in-opener-iframe.html", "_blank");
}

</script>
<body onload="setTimeout('runTest();', 0);">
This test sees if starting with an empty iframe in the markup, dynamically adding form elements to it, and dynamically submitting the form result in an extra back/forward entry.  It is important that these steps run after the initial load is complete, hence the setTimeout() for running the test.<br>
<iframe id="testFrame" name="testFrame" src="resources/otherpage.html"></iframe>
<div id="submitform" onclick="submit();">Submit form</div>
</body>
</html>
