third_party/blink/web_tests/http/tests/navigation/parsed-iframe-dynamic-form-back-entry-user-gesture-expected.txt

This test sees if starting with an empty iframe in the markup, dynamically adding form elements to it, and dynamically submitting the form result in an extra back/forward entry. It is important that these steps run after the initial load is complete, hence the setTimeout() for running the test.

Submit form

============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/parsed-iframe-dynamic-form-back-entry-user-gesture.html
            http://127.0.0.1:8000/navigation/resources/otherpage.html (in frame "testFrame")
curr->  http://127.0.0.1:8000/navigation/parsed-iframe-dynamic-form-back-entry-user-gesture.html
            http://127.0.0.1:8000/navigation/resources/notify-done.html (in frame "testFrame")
===============================================
